moved zip code input to bio form, untested
