+++ client/components/TeacherLessonForm/TeacherLessonForm.jsx
@@ -1,15 +1,11 @@
 import { useRef, useState } from "react";
 import { Button, Form } from "react-bootstrap";
-import { getCityFromZipCode } from "../../services/zipcode";
 import SubjectInputs from "../SubjectInputs/SubjectInputs";
 
-export default function TeacherLessonForm({ setSubject, setZipCode, setStep, cityName, setCityName, stateName, setStateName }) {
+export default function TeacherLessonForm({ setSubject, setZipCode, setStep }) {
 
   const subjectInputRef = useRef();
-  const zipCodeInputRef = useRef();
   const [formErrors, setFormErrors] = useState({});
-  const [showCity, setShowCity] = useState(false);
-  const [zipCodeChecked, setZipCodeChecked] = useState(false);
   const [subjects, setSubjects] = useState([1]);
 
   console.log(subjects);
@@ -18,13 +14,6 @@ export default function TeacherLessonForm({ setSubject, setZipCode, setStep, cit
 
   const isFormInvalid = () => {
     let invalid = false;
-
-    if (zipCodeInputRef.current.value === '') {
-      setFormErrors({ ...formErrors, zipCode: 'Zip code is required'});
-      invalid = true;
-    }
-
-    if (formErrors.zipCode) invalid = true;
     
     if (subjectInputRef.current.value === '') {
       setFormErrors({ ...formErrors, subject: 'Subject is required'});
@@ -46,37 +35,8 @@ export default function TeacherLessonForm({ setSubject, setZipCode, setStep, cit
     e.preventDefault();
     if (isFormInvalid()) return;
     const formData = new FormData(e.target);
-    if (!zipCodeChecked) {
-      const zipCodeResponse = await getCityFromZipCode(formData.get('zip'));
-      if (zipCodeResponse.city && zipCodeResponse.state) {
-        setCityName(zipCodeResponse.city);
-        setStateName(zipCodeResponse.state);
-        setZipCodeChecked(true);
-      }
-      else if (zipCodeResponse.error_msg) {
-        setFormErrors({ zipCode: 'Please enter a valid zip code'});
-        return;
-      }
-    }
     setSubject(formData.get('subject'));
-    setZipCode(formData.get('zip'));
     setStep(3);
-  }
-
-  const handleEnterZipCode = async (e) => {
-    if (Number(e.target.value) && e.target.value.length === 5) {
-      const zipCodeResponse = await getCityFromZipCode(e.target.value);
-      if (zipCodeResponse.city && zipCodeResponse.state) {
-        setShowCity(true);
-        setCityName(zipCodeResponse.city);
-        setStateName(zipCodeResponse.state);
-        setZipCodeChecked(true);
-      }
-      else if (zipCodeResponse.error_msg) {
-        setFormErrors({ zipCode: 'Please enter a valid zip code'});
-        setZipCodeChecked(true);
-      }
-    }
   }
 
   const handleAddSubject = () => {
@@ -97,21 +57,6 @@ export default function TeacherLessonForm({ setSubject, setZipCode, setStep, cit
       <Button variant="outline-primary" onClick={handleAddSubject}>+ Add Another Subject</Button>
 
 
-      <Form.Group className="mb-1" controlId="zipCode">
-        <Form.Label>Zip Code</Form.Label>
-        <Form.Control type="number" placeholder="97214" name="zip" ref={zipCodeInputRef} onChange={handleChangeZipCode} onBlur={handleEnterZipCode}></Form.Control>
-        {formErrors.zipCode &&
-          <Form.Text className="text-danger">{formErrors.zipCode}</Form.Text>
-        }
-      </Form.Group>
-
-      {showCity &&
-      <div>
-        <Form.Text>
-          {cityName}, {stateName}
-        </Form.Text>
-      </div>
-      }
       <Button type="submit">Next</Button>
     </Form>
   );
